--- a/index.js
+++ b/index.js
@@ -1,28 +1,28 @@
-module.exports = vonNeumann;
+module.exports = function vonNeumann(range, dimensions) {
+  range = range || 1
+  dimensions = dimensions || 2
 
-function vonNeumann(range, dims) {
-    dims = dims || 2;
-    range = range || 1;
-    return recurse([], [], 0);
-
-    function recurse(array, temp, d) {
-        var manhattanDistance,
-            i;
-
-        if (d === dims-1) {
-            for (i = -range; i <= range; i += 1) {
-                manhattanDistance = temp.reduce(function (sum, value) { return sum + Math.abs(value); }, Math.abs(i));
-
-                if (manhattanDistance <= range && manhattanDistance !== 0) {
-                    array.push(temp.concat(i));
-                }
-            }
-        } else {
-            for (i = -range; i <= range; i += 1) {
-                recurse(array, temp.concat(i), d + 1);
-            }
-        }
-
-        return array;
+  var size = range * 2 + 1
+  var iterations = Math.pow(size, dimensions)
+  var center = (iterations - 1) / 2
+  var neighbors = []
+  for (var i = 0; i < iterations; i++) {
+    if (i !== center) {
+      var neighbor = new Array(dimensions)
+      var distance = 0
+      var remaining = i
+      for (var d = 0; d < dimensions; d++) {
+        var remainder = remaining % Math.pow(size, d + 1)
+        var value = remainder / Math.pow(size, d) - range
+        neighbor[d] = value
+        distance += Math.abs(value)
+        remaining -= remainder
+      }
+      if (distance <= range) {
+        neighbors.push(neighbor)
+      }
     }
+  }
+
+  return neighbors
 }