hi, decided to try my hand at optimizing this module after doing some stuff with `hughsk/moore`. New version is ~5x faster and passes tests:thumbsup:
```md
# old version 1000000 times
ok ~5.33 s (5 s + 328003895 ns)

# new version 1000000 times
ok ~1.01 s (1 s + 11613949 ns)
```
this is unarguably a more complicated algorithm to implement. I was actually planning on using [`semibran/delannoy`](https://github.com/semibran/delannoy) to calculate the number of cells in the desired neighborhood, which would allow the algorithm to allocate the necessary memory beforehand. however, I found that it would be implausible to add that inline and I thought it would be a bad idea to add extra dependencies :grimacing:

I'll try to trim it down and see what I can come up with. if it works out, I might be able to get an extra speed boost.