@kchapelier 2017-06-28
:+1: Thanks for the PR, it looks great ! I'm a little too busy today but I'll merge and look more closely into it tomorrow.