@kchapelier 2017-06-29
This was merged and published as 1.0.1 on npm. 🎉 

I credited you in the changelog and listed you as contributor on NPM.